docs(textbox): add playground example
