+++ skills/carbon-react/components/textbox.md
@@ -359,52 +359,6 @@ description: Carbon Textbox component props and usage examples.
 | aria-grabbed | Booleanish \| undefined | No |  | Yes | in ARIA 1.1 | Indicates an element's "grabbed" state in a drag-and-drop operation. |  |
 
 ## Examples
-### Default
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("");
-
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
-}
-```
-
-
-### Character Counter
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      inputHint="Hint text (optional)."
-      value={state}
-      onChange={setValue}
-      characterLimit={10}
-    />
-  );
-}
-```
-
-
 ### Character Counter Translations
 
 **Render**
@@ -442,285 +396,6 @@ description: Carbon Textbox component props and usage examples.
       />
     </I18nProvider>
   );
-}
-```
-
-
-### Prefix
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      prefix="prefix"
-    />
-  );
-}
-```
-
-
-### Sizes
-
-**Render**
-
-```tsx
-() => {
-  const [smallState, setSmallState] = useState("");
-  const [mediumState, setMediumState] = useState("");
-  const [largeState, setLargeState] = useState("");
-  const setValue = (
-    { target }: React.ChangeEvent<HTMLInputElement>,
-    size: string,
-  ) => {
-    if (size === "small") setSmallState(target.value);
-    else if (size === "medium") setMediumState(target.value);
-    else if (size === "large") setLargeState(target.value);
-  };
-  return (
-    <Box>
-      <Textbox
-        key={`Textbox - small`}
-        label={`Textbox - small`}
-        value={smallState}
-        size={"small"}
-        mb={2}
-        onChange={(e) => setValue(e, "small")}
-        placeholder="Textbox"
-      />
-
-      <Textbox
-        key={`Textbox - medium`}
-        label={`Textbox - medium`}
-        value={mediumState}
-        size={"medium"}
-        mb={2}
-        onChange={(e) => setValue(e, "medium")}
-        placeholder="Textbox"
-      />
-
-      <Textbox
-        key={`Textbox - large`}
-        label={`Textbox - large`}
-        value={largeState}
-        size={"large"}
-        mb={2}
-        onChange={(e) => setValue(e, "large")}
-        placeholder="Textbox"
-      />
-    </Box>
-  );
-}
-```
-
-
-### Margins
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} m={4} />;
-}
-```
-
-
-### Disabled
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      inputIcon="calendar"
-      label="Textbox"
-      disabled
-      value={state}
-      onChange={setValue}
-    />
-  );
-}
-```
-
-
-### Read Only
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      readOnly
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
-}
-```
-
-
-### With Label Inline
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      labelInline
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
-}
-```
-
-
-### With Custom Label Width And Input Width
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      labelInline
-      labelWidth={50}
-      inputWidth={50}
-    />
-  );
-}
-```
-
-
-### With Custom Max Width
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox label="Textbox" value={state} onChange={setValue} maxWidth="50%" />
-  );
-}
-```
-
-
-### With Field Help
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      fieldHelp="Help"
-    />
-  );
-}
-```
-
-
-### With Input Hint
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      inputHint="Hint text (optional)."
-    />
-  );
-}
-```
-
-
-### Required
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} required />;
-}
-```
-
-
-### IsOptional
-
-**Render**
-
-```tsx
-() => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} />;
 }
 ```
