+++ src/components/textbox/textbox.stories.tsx
@@ -4,7 +4,6 @@ import { Meta, StoryObj } from "@storybook/react-vite";
 import I18nProvider from "../i18n-provider";
 import generateStyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
-import Box from "../box";
 import Textbox from ".";
 
 const styledSystemProps = generateStyledSystemProps({
@@ -28,39 +27,33 @@ const meta: Meta<typeof Textbox> = {
 export default meta;
 type Story = StoryObj<typeof Textbox>;
 
-export const Default: Story = () => {
-  const [state, setState] = useState("");
-
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
+export const Playground: Story = {
+  render: (args) => {
+    const [state, setState] = useState("");
+    const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
+      setState(target.value);
+    };
+    return <Textbox {...args} value={state} onChange={setValue} />;
+  },
+  args: {
+    label: "Textbox",
+    placeholder: "Placeholder",
+    disabled: false,
+    readOnly: false,
+    required: false,
+    size: "medium",
+    inputHint: "Hint text",
+    prefix: "£",
+    inputIcon: "search",
+    inputWidth: 100,
+    maxWidth: "100%",
+    labelInline: false,
+    characterLimit: 50,
+    error: "",
+    warning: "",
+  },
 };
-Default.storyName = "Default";
-
-export const CharacterCounter: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      inputHint="Hint text (optional)."
-      value={state}
-      onChange={setValue}
-      characterLimit={10}
-    />
-  );
-};
-CharacterCounter.storyName = "Character Counter";
+Playground.storyName = "Playground";
 
 export const CharacterCounterTranslations: Story = () => {
   const [state, setState] = useState("Textbox");
@@ -96,211 +89,3 @@ export const CharacterCounterTranslations: Story = () => {
   );
 };
 CharacterCounterTranslations.storyName = "Character Counter Translations";
-
-export const Prefix: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      prefix="prefix"
-    />
-  );
-};
-Prefix.storyName = "Prefix";
-
-export const Sizes: Story = () => {
-  const [smallState, setSmallState] = useState("");
-  const [mediumState, setMediumState] = useState("");
-  const [largeState, setLargeState] = useState("");
-  const setValue = (
-    { target }: React.ChangeEvent<HTMLInputElement>,
-    size: string,
-  ) => {
-    if (size === "small") setSmallState(target.value);
-    else if (size === "medium") setMediumState(target.value);
-    else if (size === "large") setLargeState(target.value);
-  };
-  return (
-    <Box>
-      <Textbox
-        key={`Textbox - small`}
-        label={`Textbox - small`}
-        value={smallState}
-        size={"small"}
-        mb={2}
-        onChange={(e) => setValue(e, "small")}
-        placeholder="Textbox"
-      />
-
-      <Textbox
-        key={`Textbox - medium`}
-        label={`Textbox - medium`}
-        value={mediumState}
-        size={"medium"}
-        mb={2}
-        onChange={(e) => setValue(e, "medium")}
-        placeholder="Textbox"
-      />
-
-      <Textbox
-        key={`Textbox - large`}
-        label={`Textbox - large`}
-        value={largeState}
-        size={"large"}
-        mb={2}
-        onChange={(e) => setValue(e, "large")}
-        placeholder="Textbox"
-      />
-    </Box>
-  );
-};
-Sizes.storyName = "Sizes";
-
-export const Margins: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} m={4} />;
-};
-Margins.storyName = "Margins";
-
-export const Disabled: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      inputIcon="calendar"
-      label="Textbox"
-      disabled
-      value={state}
-      onChange={setValue}
-    />
-  );
-};
-Disabled.storyName = "Disabled";
-
-export const ReadOnly: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      readOnly
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
-};
-ReadOnly.storyName = "Read Only";
-
-export const WithLabelInline: Story = () => {
-  const [state, setState] = useState("");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return (
-    <Textbox
-      label="Textbox"
-      labelInline
-      value={state}
-      onChange={setValue}
-      placeholder="Textbox"
-    />
-  );
-};
-WithLabelInline.storyName = "With Label Inline";
-
-export const WithCustomLabelWidthAndInputWidth: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      labelInline
-      labelWidth={50}
-      inputWidth={50}
-    />
-  );
-};
-WithCustomLabelWidthAndInputWidth.storyName =
-  "With Custom Label Width And Input Width";
-
-export const WithCustomMaxWidth: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox label="Textbox" value={state} onChange={setValue} maxWidth="50%" />
-  );
-};
-WithCustomMaxWidth.storyName = "With Custom Max Width";
-
-export const WithFieldHelp: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      fieldHelp="Help"
-    />
-  );
-};
-WithFieldHelp.storyName = "With Field Help";
-
-export const WithInputHint: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-
-  return (
-    <Textbox
-      label="Textbox"
-      value={state}
-      onChange={setValue}
-      inputHint="Hint text (optional)."
-    />
-  );
-};
-WithInputHint.storyName = "With Input Hint";
-
-export const Required: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} required />;
-};
-Required.storyName = "Required";
-
-export const IsOptional: Story = () => {
-  const [state, setState] = useState("Textbox");
-  const setValue = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    setState(target.value);
-  };
-  return <Textbox label="Textbox" value={state} onChange={setValue} />;
-};
-IsOptional.storyName = "IsOptional";
